% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_amazon_alexa.R
\docType{data}
\name{amazon_alexa}
\alias{amazon_alexa}
\title{Amazon Alexa Reviews}
\format{
A data frame with 3150 rows and 5 variables:
\describe{
\item{rating}{The 1-5 rating}
\item{date}{The date of the rating}
\item{variation}{The type of Alexa}
\item{feedback}{The text feedback}
\item{verified_reviews}{Whether the review was verified (1) or not (0)}
}
}
\source{
\url{https://www.kaggle.com/sid321axn/amazon-alexa-reviews}
}
\usage{
amazon_alexa
}
\description{
This dataset contains text reviews as well as the 1-5 rating from customers who bought an Alexa device on Amazon.
}
\keyword{datasets}
