% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demo.R
\docType{data}
\name{demo}
\alias{demo}
\title{Importing Demo}
\format{
A data frame with 6 rows and 5 variables:
\describe{
\item{character}{letters}
\item{integer}{Numbers with no decimal places}
\item{double}{Numbers with decimal places}
\item{logical}{TRUE/FALSE values}
\item{date}{Dates in the format 31-Jan-21}
}
}
\source{
\url{https://psyteachr.github.io/reprores-v2/data/demo.csv}
}
\usage{
demo
}
\description{
A dataset with five different types of columns, and in five different formats (csv, tsv, xlsx, sav, and json), used to practice importing.
}
\keyword{datasets}
