% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scottish_population.R
\docType{data}
\name{scottish_population}
\alias{scottish_population}
\title{Scottish Population}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{name}{The county name}
\item{population}{The population (contains commas as thousands separator)}
}
}
\source{
\url{https://psyteachr.github.io/reprores-v3/data/scottish_population.csv}
}
\usage{
scottish_population
}
\description{
The population for each county in Scotland in 2021.
}
\keyword{datasets}
