% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sunfact2021.R
\docType{data}
\name{sunfact2021}
\alias{sunfact2021}
\title{Glasgow Sun}
\format{
A data frame with 365 rows and 3 variables:
\describe{
\item{Day}{The date in d/m/YYYY format}
\item{RiseTime}{Sunrise time in decimal hour format}
\item{SetTime}{Sunset time in decimal hour format}
}
}
\source{
\url{https://www.schoolsobservatory.org/learn/astro/nightsky/sunrs_set}
}
\usage{
sunfact2021
}
\description{
Sunrise and sunset times for every day of the year in Glasgow in 2021.
}
\keyword{datasets}
